+++ DORM/Includes/DORMInfo.php
@@ -4,6 +4,6 @@ namespace DORM\Includes;
 final class DORMInfo {
 
 	final static public function getVersion(): string {
-		return '0.1.3';
+		return '0.1.5';
 	}
 }